trading-backend/requirements.txt
fastapi==0.115.5
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
SQLAlchemy==1.4.31
uvicorn==0.32.0
pymysql
email-validator
python-multipart
